refactor: initial work on using Win10 APIs
